2-Backend/cers_backend/apps/emergencies/serializers.py
from rest_framework import serializers
from .models import Emergency

class EmergencySerializer(serializers.ModelSerializer):
    class Meta:
        model = Emergency
        fields = '__all__'
        read_only_fields = ['reporter', 'status']
